Change my web site link to go to my MIT web page.
For consistency with my listings on the SDG web site and the CSAIL web
site.  Yes, my MIT home page is not currently as nice as my personal
home page (not that that one is great either), but it's more relevant,
and I hope to improve it at some point.